requirements.txt
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.40.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
distlib==0.3.8
filelock==3.15.3
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
html2image==2.0.4.3
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
setuptools==70.1.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12.1
thop==0.1.1.post2209072238
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.36
ultralytics-thop==2.0.0
urllib3==2.2.2
virtualenv==20.26.2
websocket-client==1.8.0
